Module wallet_core::account

The account module. Interaction with an Account happens via an AccountHandle.

Modules

builder

The module with the AccountBuilder.

constants

Constants used for the account and account operations.

handle

A thread guard over an account, all account methods are called from here.

operations

The account operations like address generation, syncing and creating transfers.

types

Types used in an account and returned from methods.

Structs

Account

An Account.

AccountOptions

Account options

AddressGenerationOptions

Options for address generation

TransferOptions

Options for value transfers

TransferOutput

An output for a value transfer

Enums

RemainderValueStrategy

The strategy to use for the remainder value management when sending funds.