Module wallet_core::c

This module expose handy C compatible functions to reuse in the different C style bindings that we have (wallet-c, wallet-jni…)

Modules

fragment

settings

time

vote

Structs

Settings

SpendingCounters

TransactionOut

Constants

FRAGMENT_ID_LENGTH

NONCES_SIZE

Functions

delete_buffer⚠

Delete a binary buffer that was returned by this library alongside with its length.

spending_counters_delete⚠

Release the memory holding the spending counters

symmetric_cipher_decrypt⚠

decrypt payload of the wallet transfer protocol

wallet_confirm_transaction⚠

Confirm the previously generated transaction identified by fragment_id

wallet_delete_error⚠

delete the pointer and free the allocated memory

wallet_delete_proposal⚠

delete the pointer

wallet_delete_settings⚠

delete the pointer and free the allocated memory

wallet_delete_wallet⚠

delete the pointer, zero all the keys and free the allocated memory

wallet_id⚠

get the wallet id

wallet_import_keys⚠

recover a wallet from an account and a list of utxo keys

wallet_set_state⚠

update the wallet account state

wallet_spending_counters⚠

get the current spending counter for the (only) account in this wallet

wallet_total_value⚠

get the total value in the wallet

wallet_vote_cast⚠

build the vote cast transaction

Type Definitions

ErrorPtr

FragmentPtr

ProposalPtr

SettingsPtr

WalletPtr